+++ modulo4/projeto-labex/src/Rotas/Rotas.js
@@ -1,12 +1,12 @@
 import React from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import AdminHomePage from "../pages/AdminHomePage";
-import ApplicationFormPage from "../pages/ApplicationFormPage";
-import CreateTripPage from "../pages/CreateTripPage";
-import HomePage from "../pages/HomePage";
-import ListTripsPage from "../pages/ListTripsPage";
-import LoginPage from "../pages/LoginPage";
-import TripDetailsPage from "../pages/TripDetailsPage";
+import AdminHomePage from "../pages/AdminHomePage/AdminHomePage";
+import ApplicationFormPage from "../pages/ApplicationFormPage/ApplicationFormPage";
+import CreateTripPage from "../pages/CreateTripPage/CreateTripPage";
+import HomePage from "../pages/HomePage/HomePage";
+import ListTripsPage from "../pages/ListTripsPage/ListTripsPage";
+import LoginPage from "../pages/LoginPage/LoginPage";
+import TripDetailsPage from "../pages/TripDetailsPage/TripDetailsPage";
 
 
 const Rotas = () => {
